+++ kibot/misc.py
@@ -35,6 +35,7 @@ SVG_SCH_PRINT = 21
 CORRUPTED_SCH = 22
 WRONG_INSTALL = 23
 RENDER_3D_ERR = 24
+FAILED_EXECUTE = 25
 error_level_to_name = ['NONE',
                        'INTERNAL_ERROR',
                        'WRONG_ARGUMENTS',
@@ -60,6 +61,7 @@ error_level_to_name = ['NONE',
                        'CORRUPTED_SCH',
                        'WRONG_INSTALL',
                        'RENDER_3D_ERR',
+                       'FAILED_EXECUTE',
                        ]
 CMD_EESCHEMA_DO = 'eeschema_do'
 URL_EESCHEMA_DO = 'https://github.com/INTI-CMNB/KiAuto'
@@ -208,6 +210,7 @@ W_NOVARIANTS = '(W069) '
 W_NOENDLIB = '(W070) '
 W_NEEDSPCB = '(W071) '
 W_NOGLOBALS = '(W072) '
+W_EMPTREP = '(W073) '
 
 
 class Rect(object):